[BE/FEAT] MEMBER 몸무게 입력, 이력 조회 구현, Height, weight Double 형으로 변경

+++ backend/fns/src/main/java/ssafy/fns/domain/member/controller/dto/MemberProfileRequestDto.java
@@ -15,8 +15,8 @@ public class MemberProfileRequestDto {
 
     private String nickname;
     private Long age;
-    private Long height;
-    private Long weight;
+    private Double height;
+    private Double weight;
     private String gender;
     private Boolean isPublished;
 }
